+++ __tests__/addon-registry-project.test.js
@@ -5,13 +5,19 @@ const {
   getAddonsLoaderChain,
 } = AddonConfigurationRegistry;
 
-describe('AddonConfigurationRegistry', () => {
-  it('works in Volto', () => {
-    const base = path.join(__dirname, '..');
-    const reg = new AddonConfigurationRegistry(base);
-    expect(reg.projectRootPath).toStrictEqual(base);
-    expect(reg.addonNames).toStrictEqual(['@plone/volto-slate']);
-  });
+describe('AddonConfigurationRegistry - Project', () => {
+  jest.mock(
+    `${path.join(
+      __dirname,
+      'fixtures',
+      'test-volto-project',
+    )}/node_modules/@plone/volto/package.json`,
+    () => ({
+      // TODO: mock the packages folder inside the mocked @plone/volto to work with resolves
+      packagesFolderAddons: {},
+    }),
+    { virtual: true },
+  );
 
   it('works in a mock project directory', () => {
     const base = path.join(__dirname, 'fixtures', 'test-volto-project');